+++ crates/onboarding/src/onboarding.rs
@@ -30,6 +30,7 @@ use zed_actions::OpenOnboarding;
 mod base_keymap_picker;
 mod basics_page;
 pub mod multibuffer_hint;
+mod theme_editor;
 mod theme_preview;
 
 /// Imports settings from Visual Studio Code.